feat(admin): extract shared settings-config module for CLI and admin UI
