+++ .changeset/shared-settings-config-module.md
@@ -1,0 +1,5 @@
+---
+"nostream": minor
+---
+
+refactor: extract shared settings-config module and guided schema for admin settings editor foundation
